server.js refactor 0.1
