+++ server/routes/payments.js
@@ -1,0 +1,68 @@
+const express = require("express");
+const router = express.Router();
+const { validateRazorpayWebhook } = require("../services/paymentService");
+const { getUserIdFromEmail } = require("../services/supabaseService");
+const {
+  insertPayment,
+  getUserProfile,
+  updateUserProfile,
+} = require("../services/dbService");
+const { postToDiscordWebhook } = require("../services/webhookService");
+
+router.post("/payment-webhook", express.json(), async (req, res) => {
+  const secret = process.env.RAZORPAY_WEBHOOK_SECRET;
+  const signature = req.headers["x-razorpay-signature"];
+
+  if (validateRazorpayWebhook(req.body, signature, secret)) {
+    const event = req.body.event;
+    const { payload } = req.body;
+
+    if (event === "order.paid") {
+      // Handle the payment_link.paid event
+      const email = payload.payment?.entity?.email;
+      const amountInRs = payload.payment?.entity?.amount / 100;
+      const orderId = payload.order?.entity?.id;
+      const paymentId = payload.payment?.entity?.id;
+      const user_id = await getUserIdFromEmail(email);
+      const paymentPayload = {
+        payload,
+        user_id,
+        transaction_id: payload.payment?.entity.acquirer_data?.rrn,
+        status: "successful",
+        provider: "razorpay",
+      };
+
+      await insertPayment(paymentPayload);
+
+      const profile = await getUserProfile(user_id);
+      const { available_ships } = profile; // current
+      const profilePayload = { available_ships: available_ships + 1 }; // updated
+
+      await updateUserProfile(user_id, profilePayload);
+
+      const webhookPayload = {
+        content: "New payment received!",
+        embeds: [
+          {
+            title: "Payment Details",
+            fields: [
+              { name: "Amount", value: `Rs ${amountInRs}` },
+              { name: "Email", value: email },
+              { name: "Order ID", value: orderId },
+              { name: "Payment ID", value: paymentId },
+            ],
+          },
+        ],
+      };
+      postToDiscordWebhook(webhookPayload);
+
+      res.status(200).json({ status: "Ships added!" });
+    } else {
+      res.status(400).json({ error: "Event not handled" });
+    }
+  } else {
+    res.status(400).json({ error: "Invalid signature" });
+  }
+});
+
+module.exports = router;
